+++ assets/javascripts/kitten/components/buttons/button/stories.js
@@ -69,6 +69,8 @@ const info = {
     ~~~js
     <Button tiny>MyButton</Button>
     <Button big>MyButton</Button>
+    <Button huge>MyButton</Button>
+    <Button giant>MyButton</Button>
     ~~~
 
     #### Display
@@ -140,6 +142,8 @@ storiesOf('Buttons/Button', module)
                 <Button
                   tiny={boolean('Tiny', false)}
                   big={boolean('Big', false)}
+                  huge={boolean('Huge', false)}
+                  giant={boolean('Giant', false)}
                   fluid={boolean('Fluid', false)}
                   modifier={select('Modifier', modifierOptions, 'hydrogen')}
                   disabled={boolean('Disabled', false)}
@@ -168,6 +172,8 @@ storiesOf('Buttons/Button', module)
                 <Button
                   tiny={boolean('Tiny', false)}
                   big={boolean('Big', false)}
+                  huge={boolean('Huge', false)}
+                  giant={boolean('Giant', false)}
                   fluid={boolean('Fluid', false)}
                   modifier={select('Modifier', modifierOptions, 'hydrogen')}
                   disabled={boolean('Disabled', false)}
@@ -208,6 +214,8 @@ storiesOf('Buttons/Button', module)
                 <Button
                   tiny={boolean('Tiny', false)}
                   big={boolean('Big', false)}
+                  huge={boolean('Huge', false)}
+                  giant={boolean('Giant', false)}
                   modifier={select('Modifier', modifierOptions, 'hydrogen')}
                   fluid={boolean('Fluid', false)}
                   disabled={boolean('Disabled', false)}
@@ -236,6 +244,8 @@ storiesOf('Buttons/Button', module)
               <Button
                 tiny={boolean('Tiny', false)}
                 big={boolean('Big', false)}
+                huge={boolean('Huge', false)}
+                giant={boolean('Giant', false)}
                 modifier={select('Modifier', modifierOptions, 'helium')}
                 disabled={boolean('Disabled', false)}
               >